pkcs11: Detect the card removal from pkcs11-helper and prompt for a new PIN
This introduces a check in pkcs11-helper verifying the smart card is ready
to sign data (was not removed since last invocation). If it was, the new
code tries to ask a user for the PIN (using askpass program if defined)
and reauthenticate the card, rather than fail hard.

Signed-off: Jakub Jelen <jjelen@redhat.com
Pair-programmed-with: Nikos Mavrogiannopoulos <nmav@redhat.com>